# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FitTrack Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: number;
  trainerId: number;
  clientName: string;
  date: string;
  time: string;
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}