# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FitTrack Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  /** @minLength 1 */
  clientName: string;
  date: string;
  time: string;
  notes?: string;
}