# lib/api-zod/src/generated/types/dashboardStatsInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FitTrack Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStatsInput {
  steps?: number;
  caloriesBurned?: number;
  waterIntakeMl?: number;
  sleepHours?: number;
  heartRate?: number;
  bmiValue?: number;
  weightKg?: number;
}