# lib/api-zod/src/generated/types/goalInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FitTrack Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { GoalInputType } from './goalInputType';

export interface GoalInput {
  /** @minLength 1 */
  title: string;
  type: GoalInputType;
  targetValue: number;
  currentValue: number;
  unit: string;
  deadline?: string;
}