# lib/api-zod/src/generated/types/mealInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FitTrack Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { MealInputMealType } from './mealInputMealType';

export interface MealInput {
  /** @minLength 1 */
  name: string;
  calories: number;
  protein: number;
  carbs: number;
  fat: number;
  mealType: MealInputMealType;
}