# lib/api-zod/src/generated/types/measurementInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FitTrack Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface MeasurementInput {
  weightKg: number;
  chestCm?: number;
  waistCm?: number;
  hipsCm?: number;
}