# lib/api-zod/src/generated/types/weeklyProgress.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FitTrack Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { DayProgress } from './dayProgress';

export interface WeeklyProgress {
  days: DayProgress[];
}