# lib/api-zod/src/generated/types/workout.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FitTrack Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkoutCategory } from './workoutCategory';

export interface Workout {
  id: number;
  name: string;
  category: WorkoutCategory;
  durationMinutes: number;
  exercises: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}