# lib/api-zod/src/generated/types/workoutSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FitTrack Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkoutSummary {
  cardio: number;
  strength: number;
  yoga: number;
  hiit: number;
  totalMinutes: number;
  totalSessions: number;
}